refactor: move config.ts to config/index.ts
